+++ packages/gitlab-backend/src/service/router.test.ts
@@ -33,6 +33,30 @@ describe('createRouter', () => {
                     })
                 );
             }
+        ),
+        rest.post(
+            'https://non-existing-example.com/api/graphql',
+            (req, res, ctx) => {
+                return res(
+                    ctx.status(200),
+                    ctx.json({
+                        url: req.url.toString(),
+                        headers: req.headers.all(),
+                    })
+                );
+            }
+        ),
+        rest.post(
+            'https://non-existing-example-2.com/api/graphql',
+            (req, res, ctx) => {
+                return res(
+                    ctx.status(200),
+                    ctx.json({
+                        url: req.url.toString(),
+                        headers: req.headers.all(),
+                    })
+                );
+            }
         )
     );
 
@@ -80,7 +104,7 @@ describe('createRouter', () => {
             // this is set to let msw pass test requests through the mock server
             agent.set('User-Agent', 'supertest');
             const response = await agent.get(
-                '/api/gitlab/non-existing-example.com/projects/434'
+                '/api/gitlab/rest/non-existing-example.com/projects/434'
             );
             expect(response.status).toEqual(200);
             expect(response.body).toEqual({
@@ -99,7 +123,7 @@ describe('createRouter', () => {
             // this is set to let msw pass test requests through the mock server
             agent.set('User-Agent', 'supertest');
             const response = await agent.get(
-                '/api/gitlab/non-existing-example-2.com/projects/434'
+                '/api/gitlab/rest/non-existing-example-2.com/projects/434'
             );
             expect(response.status).toEqual(200);
             expect(response.body).toEqual({
@@ -111,6 +135,86 @@ describe('createRouter', () => {
                 },
                 url: 'https://non-existing-example-2.com/api/v4/projects/434',
             });
+        });
+    });
+
+    describe('Graphql requests', () => {
+        it('First instance should work', async () => {
+            const agent = request.agent(app);
+            // this is set to let msw pass test requests through the mock server
+            agent.set('User-Agent', 'supertest');
+            const response = await agent.post(
+                '/api/gitlab/graphql/non-existing-example.com'
+            );
+            expect(response.status).toEqual(200);
+            expect(response.body).toEqual({
+                headers: {
+                    'accept-encoding': 'gzip, deflate',
+                    connection: 'close',
+                    'content-length': '2',
+                    'content-type': 'application/json',
+                    host: 'non-existing-example.com',
+                    'user-agent': 'supertest',
+                },
+                url: 'https://non-existing-example.com/api/graphql',
+            });
+        });
+
+        it('Second instance should work', async () => {
+            const agent = request.agent(app);
+            // this is set to let msw pass test requests through the mock server
+            agent.set('User-Agent', 'supertest');
+            const response = await agent.post(
+                '/api/gitlab/graphql/non-existing-example-2.com'
+            );
+            expect(response.status).toEqual(200);
+            expect(response.body).toEqual({
+                headers: {
+                    'accept-encoding': 'gzip, deflate',
+                    connection: 'close',
+                    'content-length': '2',
+                    'content-type': 'application/json',
+                    host: 'non-existing-example-2.com',
+                    'user-agent': 'supertest',
+                },
+                url: 'https://non-existing-example-2.com/api/graphql',
+            });
+        });
+
+        it('Methods different from POST should reject', async () => {
+            const agent = request.agent(app);
+            // this is set to let msw pass test requests through the mock server
+            agent.set('User-Agent', 'supertest');
+            for (const method of [
+                // Get will be received by main server
+                'delete',
+                'put',
+                'options',
+                'trace',
+                'patch',
+            ]) {
+                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                // @ts-ignore
+                const response = await agent?.[method](
+                    '/api/gitlab/graphql/non-existing-example.com'
+                );
+
+                expect(response.status).toEqual(404);
+                expect(response.body).toEqual({});
+            }
+        });
+
+        it('Mutation queries should be rejected', async () => {
+            const agent = request.agent(app);
+            // this is set to let msw pass test requests through the mock server
+            agent.set('User-Agent', 'supertest');
+            const response = await agent
+                .post('/api/gitlab/graphql/non-existing-example-2.com')
+                .send({
+                    query: 'mutation { createIssue { id } }',
+                });
+            expect(response.status).toEqual(404);
+            expect(response.body).toEqual({});
         });
     });
 
@@ -127,9 +231,10 @@ describe('createRouter', () => {
                 'trace',
                 'patch',
             ]) {
+                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
                 // @ts-ignore
                 const response = await agent?.[method](
-                    '/api/gitlab/non-existing-example-2.com/projects/434'
+                    '/api/gitlab/rest/non-existing-example-2.com/projects/434'
                 );
                 expect(response.status).toEqual(404);
                 expect(response.body).toEqual({});
@@ -141,7 +246,7 @@ describe('createRouter', () => {
             // this is set to let msw pass test requests through the mock server
             agent.set('User-Agent', 'supertest');
             const response = await agent.get(
-                '/api/gitlab/does.not.exist/projects/434'
+                '/api/gitlab/rest/does.not.exist/projects/434'
             );
             expect(response.status).toEqual(404);
             expect(response.body).toEqual({});
@@ -166,6 +271,30 @@ describe('createRouter with baseUrl', () => {
         ),
         rest.get(
             'https://non-existing-example-2.com/api/v4/projects/434',
+            (req, res, ctx) => {
+                return res(
+                    ctx.status(200),
+                    ctx.json({
+                        url: req.url.toString(),
+                        headers: req.headers.all(),
+                    })
+                );
+            }
+        ),
+        rest.post(
+            'https://non-existing-example.com/api/graphql',
+            (req, res, ctx) => {
+                return res(
+                    ctx.status(200),
+                    ctx.json({
+                        url: req.url.toString(),
+                        headers: req.headers.all(),
+                    })
+                );
+            }
+        ),
+        rest.post(
+            'https://non-existing-example-2.com/api/graphql',
             (req, res, ctx) => {
                 return res(
                     ctx.status(200),
@@ -227,7 +356,7 @@ describe('createRouter with baseUrl', () => {
             // this is set to let msw pass test requests through the mock server
             agent.set('User-Agent', 'supertest');
             const response = await agent.get(
-                `${basePath}/api/gitlab/non-existing-example.com/projects/434`
+                `${basePath}/api/gitlab/rest/non-existing-example.com/projects/434`
             );
             expect(response.status).toEqual(200);
             expect(response.body).toEqual({
@@ -246,7 +375,7 @@ describe('createRouter with baseUrl', () => {
             // this is set to let msw pass test requests through the mock server
             agent.set('User-Agent', 'supertest');
             const response = await agent.get(
-                `${basePath}/api/gitlab/non-existing-example-2.com/projects/434`
+                `${basePath}/api/gitlab/rest/non-existing-example-2.com/projects/434`
             );
             expect(response.status).toEqual(200);
             expect(response.body).toEqual({
@@ -265,9 +394,91 @@ describe('createRouter with baseUrl', () => {
             // this is set to let msw pass test requests through the mock server
             agent.set('User-Agent', 'supertest');
             const response = await agent.get(
-                '/api/gitlab/non-existing-example.com/projects/434'
+                '/api/gitlab/rest/non-existing-example.com/projects/434'
             );
             expect(response.status).toEqual(404);
+        });
+    });
+
+    describe('Graphql requests', () => {
+        it('First instance should work', async () => {
+            const agent = request.agent(app);
+            // this is set to let msw pass test requests through the mock server
+            agent.set('User-Agent', 'supertest');
+            const response = await agent.post(
+                `${basePath}/api/gitlab/graphql/non-existing-example.com`
+            );
+            expect(response.status).toEqual(200);
+            expect(response.body).toEqual({
+                headers: {
+                    'accept-encoding': 'gzip, deflate',
+                    connection: 'close',
+                    'content-length': '2',
+                    'content-type': 'application/json',
+                    host: 'non-existing-example.com',
+                    'user-agent': 'supertest',
+                },
+                url: 'https://non-existing-example.com/api/graphql',
+            });
+        });
+
+        it('Second instance should work', async () => {
+            const agent = request.agent(app);
+            // this is set to let msw pass test requests through the mock server
+            agent.set('User-Agent', 'supertest');
+            const response = await agent.post(
+                `${basePath}/api/gitlab/graphql/non-existing-example-2.com`
+            );
+            expect(response.status).toEqual(200);
+            expect(response.body).toEqual({
+                headers: {
+                    'accept-encoding': 'gzip, deflate',
+                    connection: 'close',
+                    'content-length': '2',
+                    'content-type': 'application/json',
+                    host: 'non-existing-example-2.com',
+                    'user-agent': 'supertest',
+                },
+                url: 'https://non-existing-example-2.com/api/graphql',
+            });
+        });
+
+        it('Methods different from POST should reject', async () => {
+            const agent = request.agent(app);
+            // this is set to let msw pass test requests through the mock server
+            agent.set('User-Agent', 'supertest');
+            for (const method of [
+                // Get will be received by main server
+                'delete',
+                'put',
+                'options',
+                'trace',
+                'patch',
+            ]) {
+                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                // @ts-ignore
+                const response = await agent?.[method](
+                    `${basePath}/api/gitlab/graphql/non-existing-example.com`
+                );
+
+                expect(response.status).toEqual(404);
+                expect(response.body).toEqual({});
+            }
+        });
+
+        it('Mutation queries should be rejected', async () => {
+            const agent = request.agent(app);
+            // this is set to let msw pass test requests through the mock server
+            agent.set('User-Agent', 'supertest');
+            const response = await agent
+                .post(
+                    `${basePath}/api/gitlab/graphql/non-existing-example-2.com`
+                )
+                .send({
+                    query: 'mutation { createIssue { id } }',
+                });
+            expect(response.status).toEqual(404);
+            expect(response.body).toEqual({});
         });
     });
 
@@ -284,9 +495,10 @@ describe('createRouter with baseUrl', () => {
                 'trace',
                 'patch',
             ]) {
+                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
                 // @ts-ignore
                 const response = await agent?.[method](
-                    `${basePath}/api/gitlab/non-existing-example-2.com/projects/434`
+                    `${basePath}/api/gitlab/rest/non-existing-example-2.com/projects/434`
                 );
                 expect(response.status).toEqual(404);
                 expect(response.body).toEqual({});
@@ -298,7 +510,7 @@ describe('createRouter with baseUrl', () => {
             // this is set to let msw pass test requests through the mock server
             agent.set('User-Agent', 'supertest');
             const response = await agent.get(
-                `${basePath}/api/gitlab/does.not.exist/projects/434`
+                `${basePath}/api/gitlab/rest/does.not.exist/projects/434`
             );
             expect(response.status).toEqual(404);
             expect(response.body).toEqual({});