+++ packages/gitlab/src/api/GitlabCIApi.ts
@@ -23,6 +23,24 @@ export type GitlabCIBuilder = {
     build(gitlabInstance: string): GitlabCIApi;
 };
 
+export type GraphQLQuery = {
+    variables: Record<string, string>;
+    query: string;
+};
+
+export type GitlabProjectCoverageResponse = {
+    data: {
+        project: {
+            pipelines: {
+                nodes: {
+                    coverage: number;
+                    createdAt: string;
+                }[];
+            };
+        };
+    };
+};
+
 export type GitlabCIApi = {
     getPipelineSummary(
         projectID: string | number
@@ -42,6 +60,10 @@ export type GitlabCIApi = {
         projectID: string | number
     ): Promise<LanguagesSummary | undefined>;
     getProjectDetails(projectSlug: string): Promise<ProjectSchema | undefined>;
+    getProjectCoverage(
+        projectSlug: string,
+        projectDefaultBranch: string
+    ): Promise<GitlabProjectCoverageResponse | undefined>;
     getIssuesSummary(
         projectID: string | number
     ): Promise<IssueSchema[] | undefined>;
@@ -53,7 +75,6 @@ export type GitlabCIApi = {
     getReleasesSummary(
         projectID: string | number
     ): Promise<ReleaseSchema[] | undefined>;
-
     getContributorsLink(
         projectWebUrl: string,
         projectDefaultBranch: string
@@ -63,7 +84,6 @@ export type GitlabCIApi = {
         projectDefaultBranch: string,
         codeOwnersPath?: string
     ): string;
-
     getReadme(
         projectID: string | number,
         branch?: string,