+++ packages/gitlab/src/api/GitlabCIClient.ts
@@ -4,6 +4,8 @@ import { parseCodeOwners } from '../components/utils';
 import {
     ContributorsSummary,
     GitlabCIApi,
+    GitlabProjectCoverageResponse,
+    GraphQLQuery,
     LanguagesSummary,
 } from './GitlabCIApi';
 
@@ -17,6 +19,7 @@ import type {
     RepositoryContributorSchema,
     UserSchema,
 } from '@gitbeaker/rest';
+import dayjs from 'dayjs';
 
 export class GitlabCIClient implements GitlabCIApi {
     discoveryApi: DiscoveryApi;
@@ -70,21 +73,29 @@ export class GitlabCIClient implements GitlabCIApi {
 
     protected async callApi<T>(
         path: string,
-        query: { [key in string]: string }
+        query: { [key in string]: string },
+        APIkind: 'rest' | 'graphql' = 'rest',
+        options: RequestInit = {}
     ): Promise<T | undefined> {
-        const apiUrl = `${await this.discoveryApi.getBaseUrl('gitlab')}/${
-            this.gitlabInstance
-        }`;
+        const apiUrl = `${await this.discoveryApi.getBaseUrl(
+            'gitlab'
+        )}/${APIkind}/${this.gitlabInstance}`;
         const token = (await this.identityApi.getCredentials()).token;
-        const options = token
-            ? {
-                  headers: {
-                      Authorization: `Bearer ${token}`,
-                  },
-              }
-            : {};
+
+        if (token) {
+            options = {
+                ...options,
+                headers: {
+                    ...options?.headers,
+                    Authorization: `Bearer ${token}`,
+                },
+            };
+        }
+
         const response = await fetch(
-            `${apiUrl}/${path}?${new URLSearchParams(query).toString()}`,
+            `${apiUrl}${path ? `/${path}` : ''}?${new URLSearchParams(
+                query
+            ).toString()}`,
             options
         );
         if (response.status === 200) {
@@ -99,6 +110,18 @@ export class GitlabCIClient implements GitlabCIApi {
             }
         }
         return undefined;
+    }
+
+    protected callGraphQLApi<T>(query: GraphQLQuery): Promise<T | undefined> {
+        const options = {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify(query),
+        };
+
+        return this.callApi<T>('', {}, 'graphql', options);
     }
 
     async getPipelineSummary(
@@ -263,6 +286,40 @@ export class GitlabCIClient implements GitlabCIApi {
         );
     }
 
+    async getProjectCoverage(
+        projectSlug: string,
+        projectDefaultBranch: string
+    ): Promise<GitlabProjectCoverageResponse | undefined> {
+        if (!projectSlug) return undefined;
+
+        return this.callGraphQLApi<GitlabProjectCoverageResponse>({
+            variables: {
+                projectSlug,
+                projectDefaultBranch,
+                updatedAfter: dayjs().subtract(30, 'days').format('YYYY-MM-DD'),
+            },
+            query: /* GraphQL */ `
+                query getProjectCoverage(
+                    $projectSlug: ID!
+                    $updatedAfter: Time
+                    $projectDefaultBranch: String
+                ) {
+                    project(fullPath: $projectSlug) {
+                        pipelines(
+                            ref: $projectDefaultBranch
+                            updatedAfter: $updatedAfter
+                        ) {
+                            nodes {
+                                coverage
+                                createdAt
+                            }
+                        }
+                    }
+                }
+            `,
+        });
+    }
+
     async getCodeOwners(
         projectID: string | number,
         branch = 'HEAD',
@@ -317,7 +374,9 @@ export class GitlabCIClient implements GitlabCIApi {
         if (filePath.startsWith('./')) filePath = filePath.slice(2);
 
         const readmeStr = await this.callApi<string>(
-            `projects/${projectID}/repository/files/${encodeURIComponent(filePath)}/raw`,
+            `projects/${projectID}/repository/files/${encodeURIComponent(
+                filePath
+            )}/raw`,
             { ref: branch }
         );