+++ packages/gitlab/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.2.0-alpha.3](https://github.com/immobiliare/backstage-plugin-gitlab/compare/v6.2.0-alpha.2...v6.2.0-alpha.3) (2023-08-29)
+
+### Bug Fixes
+
+-   no data error ([b44eb1d](https://github.com/immobiliare/backstage-plugin-gitlab/commit/b44eb1de87add06bc7ff10a7c7fd69358644add9))
+
+# [6.2.0-alpha.2](https://github.com/immobiliare/backstage-plugin-gitlab/compare/v6.2.0-alpha.1...v6.2.0-alpha.2) (2023-08-28)
+
+### Bug Fixes
+
+-   path can be empty string ([1195451](https://github.com/immobiliare/backstage-plugin-gitlab/commit/1195451819930167f6c6323383a5f6b727484792))
+
+# [6.2.0-alpha.1](https://github.com/immobiliare/backstage-plugin-gitlab/compare/v6.2.0-alpha.0...v6.2.0-alpha.1) (2023-08-28)
+
+### Bug Fixes
+
+-   graphql query ([cf3aefb](https://github.com/immobiliare/backstage-plugin-gitlab/commit/cf3aefb211da684b07120428a5c4c977acef36dc))
+-   routes intersection ([ea6de8c](https://github.com/immobiliare/backstage-plugin-gitlab/commit/ea6de8cedd623face5153473d68305dbb5282963))
+
+# [6.2.0-alpha.0](https://github.com/immobiliare/backstage-plugin-gitlab/compare/v6.1.0...v6.2.0-alpha.0) (2023-08-25)
+
+### Bug Fixes
+
+-   Fix prettier formatting ([fc8f4c0](https://github.com/immobiliare/backstage-plugin-gitlab/commit/fc8f4c0ad1cce7a8efaef45070a6ff05f2a453df))
+-   Fix review comments ([5cf88d3](https://github.com/immobiliare/backstage-plugin-gitlab/commit/5cf88d3148aa16461bb62c0965d652517fc5db14))
+
+### Features
+
+-   Add Coverage card ([0946e52](https://github.com/immobiliare/backstage-plugin-gitlab/commit/0946e529623bd578e21831bbdbbfcb1072ac739c))
+-   Add graphql query to mock ([e035819](https://github.com/immobiliare/backstage-plugin-gitlab/commit/e035819ade4aeec500e4571b940dca1f44306b75))
+-   Add info card variant prop ([aed18ec](https://github.com/immobiliare/backstage-plugin-gitlab/commit/aed18ecc0021a2e89b311d172bafd8833b7760ac))
+-   Add tests, refactored gitlab client ([8577210](https://github.com/immobiliare/backstage-plugin-gitlab/commit/8577210eecd4f9271672eab3db5d375705107cf2))
+
 # [6.1.0](https://github.com/immobiliare/backstage-plugin-gitlab/compare/v6.1.0-alpha.0...v6.1.0) (2023-08-22)
 
 **Note:** Version bump only for package @immobiliarelabs/backstage-plugin-gitlab